refactor(presentation/placedetails): create component files
