+++ src/presentation/placeDetails/components/index.ts
@@ -1,0 +1,3 @@
+export { default as GallerySummaryImages } from './gallerySummaryImages';
+export { default as MoreDetails } from './moreDetails';
+export { default as Reviews } from './reviews';
