+++ tests/presentation/placeDetails/index.spec.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react-native';
 import { faker } from '@faker-js/faker';
-import PlaceDetails, {
-  getStyleOfPhotoOfReviewProfile,
-} from '../../../src/presentation/placeDetails';
+import { getStyleOfPhotoOfReviewProfile } from '../../../src/presentation/placeDetails/components/reviews';
+import PlaceDetails from '../../../src/presentation/placeDetails';
 
 describe('PlaceDetails: getStyleOfPhotoOfReviewProfile', () => {
   test('should get style of photo of review profile when index equals 0', () => {
